Adiciona X-XRSF-Token no header de requisições que não sejam GET
Refatora CadastroVeiculo par receber validacoes do utils.
