+++ src/services/AutenticacaoService.js
@@ -2,6 +2,7 @@ const baseUrl = 'http://localhost:8080/auth';
 
 const headers = new Headers({
   'Content-Type': 'application/json',
+  'X-XSRF-TOKEN': '',
   Authorization: 'Bearer ',
 });
