+++ src/services/UsuarioService.js
@@ -4,11 +4,14 @@ const baseUrl = 'http://localhost:8080/usuarios';
 const headers = () =>
   new Headers({
     'Content-Type': 'application/json',
+    'X-XSRF-TOKEN': '',
     Authorization: `Bearer ${localStorage.getItem(TOKEN_KEY)}`,
   });
 const UsuarioService = {
   listar() {
-    return fetch(baseUrl, { headers: headers() }).then((response) =>
+    const reqHeaders = headers();
+    reqHeaders.delete('X-XSRF-TOKEN');
+    return fetch(baseUrl, { headers: reqHeaders }).then((response) =>
       response.json()
     );
   },
